refactor(whispering): invert ShortcutTable to a row snippet
ShortcutTable took a type: 'local' | 'global' discriminator and re-demuxed it
internally three times (default lookup, pressedKeys creation, recorder choice),
even though the page already branches on tauri. The two modes share only the
search-and-table chrome; their data flow is disjoint.

Genericize the table to platform-agnostic chrome with a row snippet, and move
the per-mode wiring up to the page that already knows the platform: each branch
owns its default lookup and recorder. createPressedKeys (a window keydown
listener that preventDefaults every key) is now created only on the browser
path that consumes it, instead of being instantiated and unused on desktop.

Removes the type/tauri props and the internal branches.
